move typescript compile to server folder
